Add files via upload
This is the original version of this application. It has a login page, a registration page, the logic to check the database for the email and password, and a form to add users to the application. This upload also contains the SQL to build the database on MSSQL. Please ensure a user is addedd and given db_owner rights to the database and schema.